fix: id to arn for private subnet
We use arn for all other outputs that return arns except this one. It makes sense to standardize it. I also cleaned up a duplicate -- but less well developed -- output with the arn name created because I was myself confused by this 😂

Note that this would be a breaking change and should only be included in a major version release.